+++ src/Symfony/Bundle/FrameworkBundle/Command/AboutCommand.php
@@ -23,6 +23,8 @@ use Symfony\Component\HttpKernel\KernelInterface;
  * A console command to display information about the current installation.
  *
  * @author Roland Franssen <franssen.roland@gmail.com>
+ *
+ * @final since version 3.4
  */
 class AboutCommand extends ContainerAwareCommand
 {
@@ -45,7 +47,7 @@ class AboutCommand extends ContainerAwareCommand
         $io = new SymfonyStyle($input, $output);
 
         /** @var $kernel KernelInterface */
-        $kernel = $this->getContainer()->get('kernel');
+        $kernel = $this->getApplication()->getKernel();
 
         $io->table(array(), array(
             array('<info>Symfony</>'),