+++ src/Symfony/Bundle/FrameworkBundle/Command/CacheWarmupCommand.php
@@ -15,14 +15,37 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
+use Symfony\Component\HttpKernel\CacheWarmer\CacheWarmerAggregate;
 
 /**
  * Warmup the cache.
  *
  * @author Fabien Potencier <fabien@symfony.com>
+ *
+ * @final since version 3.4
  */
 class CacheWarmupCommand extends ContainerAwareCommand
 {
+    private $cacheWarmer;
+
+    /**
+     * @param CacheWarmerAggregate $cacheWarmer
+     */
+    public function __construct($cacheWarmer = null)
+    {
+        parent::__construct();
+
+        if (!$cacheWarmer instanceof CacheWarmerAggregate) {
+            @trigger_error(sprintf('Passing a command name as the first argument of "%s" is deprecated since version 3.4 and will be removed in 4.0. If the command was registered by convention, make it a service instead.', __METHOD__), E_USER_DEPRECATED);
+
+            $this->setName(null === $cacheWarmer ? 'cache:warmup' : $cacheWarmer);
+
+            return;
+        }
+
+        $this->cacheWarmer = $cacheWarmer;
+    }
+
     /**
      * {@inheritdoc}
      */
@@ -54,18 +77,22 @@ EOF
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        $io = new SymfonyStyle($input, $output);
-
-        $kernel = $this->getContainer()->get('kernel');
-        $io->comment(sprintf('Warming up the cache for the <info>%s</info> environment with debug <info>%s</info>', $kernel->getEnvironment(), var_export($kernel->isDebug(), true)));
-
-        $warmer = $this->getContainer()->get('cache_warmer');
-
-        if (!$input->getOption('no-optional-warmers')) {
-            $warmer->enableOptionalWarmers();
+        // BC to be removed in 4.0
+        if (null === $this->cacheWarmer) {
+            $this->cacheWarmer = $this->getContainer()->get('cache_warmer');
+            $cacheDir = $this->getContainer()->getParameter('kernel.cache_dir');
         }
 
-        $warmer->warmUp($this->getContainer()->getParameter('kernel.cache_dir'));
+        $io = new SymfonyStyle($input, $output);
+
+        $kernel = $this->getApplication()->getKernel();
+        $io->comment(sprintf('Warming up the cache for the <info>%s</info> environment with debug <info>%s</info>', $kernel->getEnvironment(), var_export($kernel->isDebug(), true)));
+
+        if (!$input->getOption('no-optional-warmers')) {
+            $this->cacheWarmer->enableOptionalWarmers();
+        }
+
+        $this->cacheWarmer->warmUp(isset($cacheDir) ? $cacheDir : $kernel->getContainer()->getParameter('kernel.cache_dir'));
 
         $io->success(sprintf('Cache for the "%s" environment (debug=%s) was successfully warmed.', $kernel->getEnvironment(), var_export($kernel->isDebug(), true)));
     }