+++ src/Symfony/Bundle/FrameworkBundle/Command/TranslationDebugCommand.php
@@ -17,24 +17,54 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\HttpKernel\Kernel;
+use Symfony\Component\HttpKernel\KernelInterface;
 use Symfony\Component\Translation\Catalogue\MergeOperation;
+use Symfony\Component\Translation\Extractor\ExtractorInterface;
 use Symfony\Component\Translation\MessageCatalogue;
 use Symfony\Component\Translation\Translator;
 use Symfony\Component\Translation\DataCollectorTranslator;
 use Symfony\Component\Translation\LoggingTranslator;
+use Symfony\Component\Translation\TranslatorInterface;
 
 /**
  * Helps finding unused or missing translation messages in a given locale
  * and comparing them with the fallback ones.
  *
  * @author Florian Voutzinos <florian@voutzinos.com>
+ *
+ * @final since version 3.4
  */
 class TranslationDebugCommand extends ContainerAwareCommand
 {
+    private $translator;
+    private $loader;
+    private $extractor;
+
     const MESSAGE_MISSING = 0;
     const MESSAGE_UNUSED = 1;
     const MESSAGE_EQUALS_FALLBACK = 2;
+
+    /**
+     * @param TranslatorInterface $translator
+     * @param TranslationLoader   $loader
+     * @param ExtractorInterface  $extractor
+     */
+    public function __construct($translator = null, TranslationLoader $loader = null, ExtractorInterface $extractor = null)
+    {
+        parent::__construct();
+
+        if (!$translator instanceof TranslatorInterface) {
+            @trigger_error(sprintf('Passing a command name as the first argument of "%s" is deprecated since version 3.4 and will be removed in 4.0. If the command was registered by convention, make it a service instead.', __METHOD__), E_USER_DEPRECATED);
+
+            $this->setName(null === $translator ? 'debug:translation' : $translator);
+
+            return;
+        }
+
+        $this->translator = $translator;
+        $this->loader = $loader;
+        $this->extractor = $extractor;
+    }
 
     /**
      * {@inheritdoc}
@@ -88,9 +118,14 @@ EOF
 
     /**
      * {@inheritdoc}
+     *
+     * BC to be removed in 4.0
      */
     public function isEnabled()
     {
+        if (null !== $this->translator) {
+            return parent::isEnabled();
+        }
         if (!class_exists('Symfony\Component\Translation\Translator')) {
             return false;
         }
@@ -103,14 +138,19 @@ EOF
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
+        // BC to be removed in 4.0
+        if (null === $this->translator) {
+            $this->translator = $this->getContainer()->get('translator');
+            $this->loader = $this->getContainer()->get('translation.loader');
+            $this->extractor = $this->getContainer()->get('translation.extractor');
+        }
+
         $io = new SymfonyStyle($input, $output);
 
         $locale = $input->getArgument('locale');
         $domain = $input->getOption('domain');
-        /** @var TranslationLoader $loader */
-        $loader = $this->getContainer()->get('translation.loader');
-        /** @var Kernel $kernel */
-        $kernel = $this->getContainer()->get('kernel');
+        /** @var KernelInterface $kernel */
+        $kernel = $this->getApplication()->getKernel();
 
         // Define Root Path to App folder
         $transPaths = array($kernel->getRootDir().'/Resources/');
@@ -142,7 +182,7 @@ EOF
         $extractedCatalogue = $this->extractMessages($locale, $transPaths);
 
         // Load defined messages
-        $currentCatalogue = $this->loadCurrentMessages($locale, $transPaths, $loader);
+        $currentCatalogue = $this->loadCurrentMessages($locale, $transPaths);
 
         // Merge defined and extracted messages to get all message ids
         $mergeOperation = new MergeOperation($extractedCatalogue, $currentCatalogue);
@@ -165,7 +205,7 @@ EOF
         }
 
         // Load the fallback catalogues
-        $fallbackCatalogues = $this->loadFallbackCatalogues($locale, $transPaths, $loader);
+        $fallbackCatalogues = $this->loadFallbackCatalogues($locale, $transPaths);
 
         // Display header line
         $headers = array('State', 'Domain', 'Id', sprintf('Message Preview (%s)', $locale));
@@ -271,7 +311,7 @@ EOF
         foreach ($transPaths as $path) {
             $path = $path.'views';
             if (is_dir($path)) {
-                $this->getContainer()->get('translation.extractor')->extract($path, $extractedCatalogue);
+                $this->extractor->extract($path, $extractedCatalogue);
             }
         }
 
@@ -279,19 +319,18 @@ EOF
     }
 
     /**
-     * @param string            $locale
-     * @param array             $transPaths
-     * @param TranslationLoader $loader
+     * @param string $locale
+     * @param array  $transPaths
      *
      * @return MessageCatalogue
      */
-    private function loadCurrentMessages($locale, $transPaths, TranslationLoader $loader)
+    private function loadCurrentMessages($locale, $transPaths)
     {
         $currentCatalogue = new MessageCatalogue($locale);
         foreach ($transPaths as $path) {
             $path = $path.'translations';
             if (is_dir($path)) {
-                $loader->loadMessages($path, $currentCatalogue);
+                $this->loader->loadMessages($path, $currentCatalogue);
             }
         }
 
@@ -299,18 +338,16 @@ EOF
     }
 
     /**
-     * @param string            $locale
-     * @param array             $transPaths
-     * @param TranslationLoader $loader
+     * @param string $locale
+     * @param array  $transPaths
      *
      * @return MessageCatalogue[]
      */
-    private function loadFallbackCatalogues($locale, $transPaths, TranslationLoader $loader)
+    private function loadFallbackCatalogues($locale, $transPaths)
     {
         $fallbackCatalogues = array();
-        $translator = $this->getContainer()->get('translator');
-        if ($translator instanceof Translator || $translator instanceof DataCollectorTranslator || $translator instanceof LoggingTranslator) {
-            foreach ($translator->getFallbackLocales() as $fallbackLocale) {
+        if ($this->translator instanceof Translator || $this->translator instanceof DataCollectorTranslator || $this->translator instanceof LoggingTranslator) {
+            foreach ($this->translator->getFallbackLocales() as $fallbackLocale) {
                 if ($fallbackLocale === $locale) {
                     continue;
                 }
@@ -319,7 +356,7 @@ EOF
                 foreach ($transPaths as $path) {
                     $path = $path.'translations';
                     if (is_dir($path)) {
-                        $loader->loadMessages($path, $fallbackCatalogue);
+                        $this->loader->loadMessages($path, $fallbackCatalogue);
                     }
                 }
                 $fallbackCatalogues[] = $fallbackCatalogue;