+++ src/Symfony/Bundle/FrameworkBundle/Command/YamlLintCommand.php
@@ -21,6 +21,8 @@ use Symfony\Component\Yaml\Command\LintCommand as BaseLintCommand;
  *
  * @author Grégoire Pineau <lyrixx@lyrixx.info>
  * @author Robin Chalas <robin.chalas@gmail.com>
+ *
+ * @final since version 3.4
  */
 class YamlLintCommand extends Command
 {