+++ src/Symfony/Component/Workflow/MarkingStore/SingleStateMarkingStore.php
@@ -59,4 +59,12 @@ class SingleStateMarkingStore implements MarkingStoreInterface
     {
         $this->propertyAccessor->setValue($subject, $this->property, key($marking->getPlaces()));
     }
+
+    /**
+     * @return string
+     */
+    public function getProperty()
+    {
+        return $this->property;
+    }
 }