+++ src/Symfony/Component/Workflow/SupportStrategy/ClassInstanceSupportStrategy.php
@@ -26,4 +26,12 @@ final class ClassInstanceSupportStrategy implements SupportStrategyInterface
     {
         return $subject instanceof $this->className;
     }
+
+    /**
+     * @return string
+     */
+    public function getClassName()
+    {
+        return $this->className;
+    }
 }